* tried to clean tings up in 256 color land... small font needs work

+++ lmc/xingmp3/src/xinglmc.cpp
@@ -205,7 +205,7 @@ Error XingLMC::ExtractMediaInfo(MediaInfoEvent **pMIE) {
 								   head.prot,
 								   head.emphasis,
 								   head.mode,
-								   head.mode_ext);
+								   (float)head.mode_ext);
 			if (mie) {
 				(*pMIE)->AddChildEvent((Event *)mie);
 				mie = NULL;
